- Moved single player gameloop to its own file - Added a few unit sprites to the map
