+++ castle_quest/src/main.rs
@@ -3,11 +3,8 @@ extern crate sdl2;
 const TITLE: &str = "Castle Quest";
 const CAM_W: u32 = 1280;
 const CAM_H: u32 = 720;
-const TILE_SIZE: u32 = 32;
+pub const TILE_SIZE: u32 = 32;
 
-use sdl2::event::Event;
-use sdl2::image::LoadTexture;
-use sdl2::keyboard::Keycode;
 use sdl2::rect::Rect;
 
 #[macro_use] mod sdl_macros;
@@ -15,8 +12,7 @@ use sdl2::rect::Rect;
 mod credits;
 mod main_menu;
 mod pixel_coordinates;
-
-use pixel_coordinates::PixelCoordinates;
+mod single_player;
 
 pub enum GameState {
 	MainMenu,
@@ -69,7 +65,7 @@ fn runner(vsync:bool) {
 fn run_game_state(core: &mut SDLCore, game_state: &GameState) -> Result<GameState, String> {
 	let next_game_state = match game_state {
 		GameState::MainMenu => main_menu::main_menu(core)?,
-		GameState::SinglePlayer => run_single_player(core)?,
+		GameState::SinglePlayer => single_player::single_player(core)?,
 		GameState::Credits => credits::credits(core)?,
 		GameState::Quit => GameState::Quit,
 		_ => return Err("Invalid game state".to_string()),
@@ -116,45 +112,6 @@ fn init_sdl_core(vsync:bool) -> Result<SDLCore, String> {
 	};
 
 	Ok(core)
-}
-
-fn run_single_player(core: &mut SDLCore) -> Result<GameState, String> {
-	//Basic mock map, 48x48 2d vector filled with 1s
-	let mut map: Vec<Vec<u32>> = vec![vec![1; 48]; 48];
-	let map_width = map[0].len();
-	let map_height = map.len();
-
-	'gameloop: loop {
-		core.wincan.clear();
-
-		for event in core.event_pump.poll_iter() {
-			match event {
-				Event::Quit{..} | Event::KeyDown{keycode: Some(Keycode::Escape), ..} => break 'gameloop,
-				_ => {},
-			}
-		}
-
-		//Draw tiles & sprites
-		for i in 0..map_height {
-			for j in 0..map_width {
-				let pixel_location = PixelCoordinates::from_matrix_indices(i as u32, j as u32);
-				let dest = Rect::new(pixel_location.x as i32, pixel_location.y as i32, TILE_SIZE, TILE_SIZE);
-
-				//Draw map tile at this coordinate
-				let map_tile_texture = match map[i][j] {
-					1 => core.texture_creator.load_texture("images/grass_tile.png")?,
-					_ => return Err("Invalid map tile id".to_string()),
-				};
-
-				core.wincan.copy(&map_tile_texture, None, dest)?; 
-			}
-		}
-
-		core.wincan.present();
-	}
-
-	//Single player finished running cleanly, automatically quit game
-	Ok(GameState::Quit)
 }
 
 fn main() {
